Plugins return home
Trying to keep the modules as plugins in npm was very cumbersome. Minor changes required constant version updates. And the result was a base code disappeared and with difficulties to keep updated. Centralizing in the same repository the base code and the standard modules makes development faster. Anyway, the import of modules through packages is still available, a good option for other developers to provide their functionality.