Changes in the way the translation files are generated
To generate a better experience and avoid mismatches between the state of the different languages ​​I generated a script that compares the files.
When run 'npm run translations' the file './i18n/generic.json' is self-generated with all strings extracted from the application. After that the own script (i18n-diff.js) is executed, which returns the missing or oversubscribed translations in each language located in the i18n/translations folder.
Throughout the previous commits I was removing and correcting mistakes of typing and lost translations according to what different people were reporting.