chore(delete-nodes): separate into its own plugin
Currently each plugin match one to one with a menu item.
We want one menu item for deleting nodes, so a plugin for it
is created. In the future it will be usefull to keep related
functionality in the same plugin, and allow them to have more
than one menu item associated.