fix(tests): fix query error handling
Using queries you can use `isError` to show an error message on the ui. For some reason, with react query update, the error goes until the test causing a failure.

I tried multiple solutions:

- Using react [error boundaries](https://reactjs.org/docs/error-boundaries.html#introducing-error-boundaries) supported on queries adding `useErrorBoundary: true` to the query options
- I tried to add a query client to the rendering and [disable the error messages](https://tanstack.com/query/v4/docs/guides/testing#turn-off-network-error-logging)
- Also. I tried to mock and catch the error using Jest expectations or try catch blocks inside the test. As show on [this](https://stackoverflow.com/a/46155381/4708664) answers.

Anything of this worked. So I decided to modify code outside the test and cheat a little bit improving the test behaviour to handle the exception thrown.